requirements.txt
telethon==1.34.0
google-genai
google-generativeai
aiohttp
python-dotenv
aiofiles
aiosqlite
Pillow
huggingface-hub
